Fixed printf format

+++ lab_01_1_2/main.c
@@ -5,7 +5,7 @@
 
 int main()
 {
-    double a, b, alpha, h;
+    float a, b, alpha, h;
 
     setbuf(stdout, NULL);
     printf("Enter a b alpha:\n");
@@ -19,7 +19,6 @@ int main()
         h = -h;
     }
 
-    printf("Square is %0.2f\n", (h * (a + b) / 2.0));
+    printf("Square is %0.2lf\n", (h * (a + b) / 2.0));
     return 0;
 }
-
